Ændret en variable fra short til int

+++ EKG_Project/Files/Producer_Consumer/ECGReadingProducer.cs
@@ -47,7 +47,7 @@ public class ECGReadingProducer
     {
         while (!ct.IsCancellationRequested)
         {
-            short v = _sensor.ReadRawSample();
+            int v = _sensor.ReadRawSample();
 
             var sample = new ECGSample()
             {
